+++ car_management_backend-main/src/main/java/org/infosys/repository/CarRepository.java
@@ -14,12 +14,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {
 
 	
 	Optional<Car> findByRegistrationNumber(String registrationNumber);
 
-	Optional<Car> findByCarId(int carId);
+	Optional<Car> findByCarId(Long carId);
 	
 	List<Car> findAllByCurrentStatus(String currentStatus);
